+++ packages/gov-ui-kit/CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 -   Update interface for `Accordion.Container` to expose value prop
 -   Update styles on `Tabs.List` for latest spec
+-   Rename `indicator` property of `<Progress />` core component to `thresholdIndicator`
 
 ## [1.0.39] - 2024-07-16